In order to support template param deduction, composable_factory has the required outer and inner params first, with the variadic ctor params being last. However, this doesn't match the WinRT requirement of putting those ctor params first.

To preserve the template param deduction in the factory, we simply reorder the params to match things up.

Related work items: #13780572

+++ cppwinrt/source/sdk.h
@@ -170,7 +170,7 @@ namespace cppwinrt
         DWORD index{ 0 };
         ULONG version_parts[4]{ 0 };
         std::array<wchar_t, MAX_PATH> kit_root;
-        while (RegEnumKeyW(key.get(), index++, kit_root.data(), kit_root.size()) == ERROR_SUCCESS)
+        while (RegEnumKeyW(key.get(), index++, kit_root.data(), static_cast<DWORD>(kit_root.size())) == ERROR_SUCCESS)
         {
             if (!std::regex_search(kit_root.data(), versionMatch, versionRx))
             {
